Refactor/ImageUploadContainer 구조 변경
