+++ src/components/Icon/index.stories.jsx
@@ -20,6 +20,7 @@ export default {
           'userLine',
           'alarmLine',
           'logout',
+          'fileGray',
         ],
       },
       defaultValue: 'listFill',
