+++ src/components/ImageUploadContainer/index.stories.jsx
