+++ src/feature/ttabong/index.js
@@ -1,5 +1,4 @@
 export { default as CheckIcon } from './CheckIcon';
-export { default as ImageUploadContainer } from './ImageUploadContainer';
 export { default as LabelList } from './LabelList';
 export { default as ReasonContainer } from './ReasonContainer';
 export { default as TTaBongModal } from './TTaBongModal';
